Show a usable rate and jump pacing to the sustained interval
The rate tile floored sub-1/min values to zero and showed a dash exactly
when the number matters, and it sampled a 30s window that a multi-minute
stall swamps. Average over the whole run instead, render below 1/min as
per hour, and put the implied time-to-finish next to the progress bar.

Pacing creeped 50% per 429 while each overshoot cost minutes of stall, so
back off to whatever interval the run has actually sustained and raise
the ceiling to 90s so it can reach it.

+++ app/src/main/java/com/noctra/scout/ScraperState.kt
@@ -13,7 +13,8 @@ data class Stats(
     val status: String = "Idle",
     val cursor: Long = 0L,
     val total: Long = 0L,
-    val ratePerMin: Int = 0,
+    /** Average checks per minute across the whole run, stalls included. */
+    val ratePerMin: Double = 0.0,
     /** The delay currently in use, which auto-pacing may have raised above the setting. */
     val paceMs: Int = 0,
     /** How many times Discord has rate limited this run. */
